sf: Allow SystemServer to set the display brightness
As part of the boot up sequence, SystemServer initializes the
DisplayManagerService to manage the global lifecycle of all displays.
Currently, the DisplayManagerService is not permitted to adjust the
display brightness settings and this is causing boot up failures.

This change allows SystemServer or AID to set the display brightness
to avoid such issues.

Bug: 248249039
Change-Id: Ia57c5fc8f3875da4f5ab31ab68d24571a5599fe4